requirements.txt
asgiref==3.11.1
click==8.3.2
colorama==0.4.6
cssbeautifier==1.15.4
django==6.0.4
django-cleanup==9.0.0
django-environ==0.13.0
django-markdownx==4.0.9
djlint==1.36.4
editorconfig==0.17.1
jsbeautifier==1.15.4
json5==0.14.0
markdown==3.10.2
pathspec==1.0.4
pillow==12.2.0
pyyaml==6.0.3
regex==2026.4.4
ruff==0.15.9
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
